refactor(scancode): Factor out a `ScanCodeLicenseDataDirReader()`
Move the code for reading data from a ScanCode license data directory to
a dedicated class, for separation of concerns and to prepare for
enhancements in upcoming changes.
